requirements.txt
Django==4.2.28
djangorestframework==3.15.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
cryptography==42.0.5
python-dotenv==1.0.1
